refactor(expand-code): :recycle: refactor whole code to increase readibilty
