+++ github/enum/GitHubURL.ts
@@ -1,0 +1,5 @@
+export enum GitHubURLEnum {
+  PREFIX = "blob/",
+  HOST = "github.com",
+  RAW_HOST = "raw.githubusercontent.com",
+}
