--- github/enum/URLmodifications.ts
@@ -1,5 +1,0 @@
-export enum URLEnums {
-    REPLACE_PREFIX = "blob/",
-    REPLACE_HOST = "github.com",
-    NEW_HOST = "raw.githubusercontent.com",
-  }
